% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{aggregate_locations_by_position}
\alias{aggregate_locations_by_position}
\title{Aggregate Locations by Position
TODO: Make faster?
WARNING: build_location_description expects singular type strings.}
\usage{
aggregate_locations_by_position(dt, sep = ". ")
}
\description{
Aggregate Locations by Position
TODO: Make faster?
WARNING: build_location_description expects singular type strings.
}
\seealso{
Other location import functions: \code{\link{apply_ff_type_matches}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}

