% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{apply_ff_type_matches}
\alias{apply_ff_type_matches}
\title{Apply Type Matches}
\usage{
apply_ff_type_matches(dt, types, match_table, drop = FALSE)
}
\description{
Apply Type Matches
}
\seealso{
Other location import functions: \code{\link{aggregate_locations_by_position}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}

