% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{apply_match_table}
\alias{apply_match_table}
\title{Apply Match Table to Locations}
\usage{
apply_match_table(dt, match_table, drop = FALSE,
  types = get_ff_types(pending = FALSE))
}
\arguments{
\item{dt}{Locations data.}

\item{match_table}{Type assignments, as returned by \code{\link{get_ff_types}}.}

\item{drop}{Whether to drop unassigned rows in \code{dt}.}

\item{Falling}{Fruit types, as returned by \code{\link{get_ff_types}}.}
}
\description{
Apply Match Table to Locations
}
\seealso{
Other location import functions: \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{merge_overlapping_locations}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}
