% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{build_col_scientific_name}
\alias{build_col_scientific_name}
\title{Build Catalogue of Life (COL) Scientific Name}
\usage{
build_col_scientific_name(json)
}
\arguments{
\item{json}{(list) Results element from \code{\link{get_col_page}}.}
}
\description{
Build Catalogue of Life (COL) Scientific Name
}
\seealso{
Other COL functions: \code{\link{get_col_page}},
  \code{\link{get_col_search}},
  \code{\link{parse_col_page}},
  \code{\link{parse_col_search}}
}
