% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{build_location_description}
\alias{build_location_description}
\title{Build Location Description}
\usage{
build_location_description(type_strings, notes = NULL, sep = ". ",
  frequency_in = "[]")
}
\description{
Builds a description from its inputs:
[nx] type string, [yx] type string, ... + sep + notes (those unique and equal for all)
}
\details{
WARNING: Not splitting at commas to support types with commas in them, so type_strings need to be single-type.
}
\examples{
build_location_description(c("Apple", "Pear", "Pear"), notes = list(c("Planted 1999", "Height 10 m"), c("Planted 1999", "Height 20 m"), c("Planted 1999", "Height 30 m")))
}
\seealso{
Other Falling Fruit functions: \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}

