% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{build_location_descriptions}
\alias{build_location_descriptions}
\title{Build Location Descriptions}
\usage{
build_location_descriptions(type_strings, notes = NULL, note_sep = ". ",
  frequency = TRUE, frequency_in = "[]")
}
\arguments{
\item{type_strings}{Character vector of type strings.}

\item{notes}{List of character vectors, each of the same length.}

\item{note_sep}{Character string to seperate each element in \code{notes}.}

\item{frequency}{Whether to display frequency before \code{type_strings}.}

\item{frequency_in}{String of two characters in which to display the frequencies.}
}
\value{
Character string of the summarized and concatenated values.
}
\description{
A faster, vectorized alternative of \code{\link{build_location_description}} for single-type locations.
}
\examples{
type_strings <- c("Apple", "Pear", "Pear")
notes <- list(c("Planted 1999", "Height 10 m"), c("Planted 1999", "Height 20 m"), c("Planted 1999", "Height 20 m"))
build_location_descriptions(type_strings, notes)
notes <- list(c("Planted 1999", NA), c(NA, "Height 20 m"), c(NA, NA))
build_location_descriptions(type_strings, notes)
build_location_descriptions(type_strings, notes, frequency = FALSE)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}
