% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{build_match_table}
\alias{build_match_table}
\title{Built Match Table from Name Matches}
\usage{
build_match_table(matches, dt, types = get_ff_types(pending = FALSE),
  saved_table = NULL)
}
\arguments{
\item{matches}{Result of \code{\link{match_names_to_ff_types}}.}

\item{dt}{Original data.}

\item{types}{Falling Fruit types, as returned by \code{\link{get_ff_types}}.}

\item{saved_table}{Previous result with saved edits.}
}
\description{
TODO: Support ordered list of name fields.
}
\seealso{
Other location import functions: \code{\link{aggregate_locations_by_position}},
  \code{\link{apply_ff_type_matches}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}

