% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{build_scientific_name_table}
\alias{build_scientific_name_table}
\title{Build table of scientific names}
\usage{
build_scientific_name_table(responses, ff_type)
}
\arguments{
\item{responses}{(response list) Result of \code{\link{query_sources_about_type}}.}

\item{ff_type}{(named list) Row of \code{\link{get_ff_types}}.}
}
\description{
Build table of scientific names
}
\seealso{
Other translation functions: \code{\link{build_common_name_table}},
  \code{\link{normalize_common_name}},
  \code{\link{parse_sources_about_type}},
  \code{\link{query_sources_about_type}}
}
