% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{build_type_strings}
\alias{build_type_strings}
\title{Build Type Strings}
\usage{
build_type_strings(ids = NULL, common_names = NULL,
  scientific_names = NULL, science_in = "[]")
}
\description{
Build Type Strings
}
\examples{
build_type_strings(1, "Apple")
build_type_strings(scientific_names = "Malus domestica")
build_type_strings(scientific_names = "Malus domestica", science_in = "()")
build_type_strings(c(1, 2), c("Apple", "Pear"), c("Malus domestica", "Pyrus communis"))
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}

