% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{build_wiki_url}
\alias{build_wiki_url}
\title{Build Wiki URL}
\usage{
build_wiki_url(wiki, type, page)
}
\description{
Build Wiki URL
}
\examples{
build_wiki_url("en", "wikipedia", "Malus domestica")
build_wiki_url("commons", "wikimedia", "Malus domestica")
build_wiki_url("species", "wikimedia", "Malus domestica")
}
\seealso{
Other Wiki functions: \code{\link{get_page}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_common_names}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikispecies_common_names}}
}

