% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Categories}
\alias{Categories}
\title{Falling Fruit Categories}
\format{An object of class \code{character} of length 4.}
\usage{
Categories
}
\description{
Names of the categories returned by Falling Fruit as bitmasks.
}
\seealso{
Other constants: \code{\link{Earth_circumference}},
  \code{\link{Earth_radius}},
  \code{\link{Google_cs_languages}},
  \code{\link{Language_codes}},
  \code{\link{Location_import_fields}},
  \code{\link{Taxonomic_ranks}}, \code{\link{Unicode}}
}
\keyword{datasets}
