% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ch1903_to_wgs84}
\alias{ch1903_to_wgs84}
\title{Convert Swiss Projection CH1903 (E, N) to WGS84 (lng, lat)}
\usage{
ch1903_to_wgs84(x, y = NULL)
}
\description{
See http://www.swisstopo.admin.ch/internet/swisstopo/de/home/topics/survey/sys/refsys/switzerland.parsysrelated1.24280.downloadList.87003.DownloadFile.tmp/ch1903wgs84de.pdf (but switch x and y)
}
\examples{
ch1903_to_wgs84(0, 1000)
}
\seealso{
Other helper functions: \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{unique_na}}
}

