% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_strings}
\alias{clean_strings}
\title{Clean Strings}
\usage{
clean_strings(x)
}
\description{
Performs general string cleaning operations.
}
\examples{
clean_strings(" ..[] Hello,,  `world`.. how are you?() ")
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{unique_na}}
}

