% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websearch.R
\name{count_gigablast_results}
\alias{count_gigablast_results}
\title{Count Gigablast Search Results (DEPRECATED)}
\usage{
count_gigablast_results(string)
}
\description{
Free. No quota is specified, but requests are often blocked. See https://gigablast.com/api.html
}
\examples{
count_gigablast_results("'Malus domestica'+'Apfel'")
}
\seealso{
Other web search functions: \code{\link{count_google_cs_results}}
}

