% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websearch.R
\name{count_google_cs_results}
\alias{count_google_cs_results}
\title{Count Google Custom Search (CS) Results}
\usage{
count_google_cs_results(string, language = NULL)
}
\description{
Free up to 100/day. Maximum 10,000/day. Rate limit 1/s. See https://developers.google.com/custom-search/json-api/v1/overview.
}
\examples{
count_google_cs_results("'Malus domestica'+'Apfel'", "en")
count_google_cs_results("'Malus domestica'+'Apfel'", "de")
}
\seealso{
Other web search functions: \code{\link{count_gigablast_results}}
}

