% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Earth_radius}
\alias{Earth_radius}
\title{Earth radius}
\format{An object of class \code{numeric} of length 1.}
\usage{
Earth_radius
}
\description{
The radius of the Earth, in meters.
}
\seealso{
Other constants: \code{\link{Categories}},
  \code{\link{Earth_circumference}},
  \code{\link{Google_cs_languages}},
  \code{\link{Language_codes}},
  \code{\link{Location_import_fields}},
  \code{\link{Taxonomic_ranks}}, \code{\link{Unicode}}
}
\keyword{datasets}
