% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_list_by_split}
\alias{expand_list_by_split}
\title{Expand List by Split}
\usage{
expand_list_by_split(x, name, split = "\\\\s*,\\\\s*")
}
\description{
NOTE: Currently unused.
}
\examples{
x <- list(list(count = 1, label = "a, b"), list(count = 2, label = "c"))
str(expand_list_by_split(x, "label"))
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{unique_na}}
}

