% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_scientific_names}
\alias{format_scientific_names}
\title{Format Scientific Names}
\usage{
format_scientific_names(x, connecting_terms = TRUE, cultivars = TRUE)
}
\arguments{
\item{x}{Character vector.}

\item{connecting_terms}{Whether to retain connecting terms (e.g. "subsp.", "var.").}

\item{cultivars}{Whether to retain cultivar names (e.g. "'Bradford'").}
}
\value{
Ideally, correctly formatted scientific names.
}
\description{
Attempts to clean and format scientific names by the following convention:\cr
<Genus|Taxon> <species> (<subg|subsp|var|f|subvar|subf>. <subspecies name>) <Author Citation> 'Cultivar'
}
\examples{
format_scientific_names("Malus")
format_scientific_names("Malus Batt.")
format_scientific_names("Tamarix rubella Batt.")
format_scientific_names("Tamarix lucronensis Sennen & Elias")
format_scientific_names("Malus x domestica var. gala", connecting_terms = TRUE)
format_scientific_names("Malus x domestica var. gala", connecting_terms = FALSE)
format_scientific_names("Malus pumila 'gala'", cultivars = TRUE)
format_scientific_names("Malus pumila 'gala'", cultivars = FALSE)
format_scientific_names("Prunus subg amygdalus")
format_scientific_names("Malus x", connecting_terms = FALSE)
}
\seealso{
\code{\link{clean_strings}} for generic string cleaning.

Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}}, \code{\link{is.empty}},
  \code{\link{lvapply}}, \code{\link{melt_by_listcol}},
  \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sp_transform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
