% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_strings}
\alias{format_strings}
\title{Format Specialty Strings}
\usage{
format_strings(x, types = "", clean = TRUE)
}
\description{
Supports species common names, latin names, and addresses.
}
\examples{
format_strings(" 123 SE. MCDONALD AV ", types = "address")
format_strings("Malus X Domestica Subsp gala 'gala'", types = "printed_scientific_name")
format_strings("Malus X Domestica Subsp gala 'gala'", types = "matched_scientific_name")
format_strings("Tamarix rubella Batt.", types = "printed_scientific_name")
format_strings("Tamarix lucronensis Sennen & Elias", types = "printed_scientific_name")
format_strings("Tamarix laxa var. subspicata Ehrenb.", types = "printed_scientific_name")
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{is.empty}}, \code{\link{melt_by_listcol}},
  \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}

