% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruitr.R
\docType{package}
\name{fruitr}
\alias{fruitr}
\alias{fruitr-package}
\title{fruitr: R functions for Falling Fruit}
\description{
The fruitr package provides two categories of functions:
functions for species common name translations and functions for location data import.
}

