% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{gct_languages}
\alias{gct_languages}
\title{List supported languages}
\source{
\url{https://cloud.google.com/translate/docs/discovering-supported-languages}
}
\usage{
gct_languages(key)
}
\arguments{
\item{key}{(character) API key.}
}
\description{
Lists the language codes supported by the Google Cloud Translation API.
}
