% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{geohashes_to_gridcells}
\alias{geohashes_to_gridcells}
\title{Convert Geohashes to Grid Cell Indices}
\usage{
geohashes_to_gridcells(geohashes)
}
\arguments{
\item{geohashes}{A character vector of geohashes.}
}
\value{
Rows of Web Mercator grid cell indices (x, y, zoom).
}
\description{
Converts binary geohashes to Web Mercator (\href{http://spatialreference.org/ref/sr-org/7483/}{EPSG:3857}) quadtree grid cell indices.
}
\examples{
geohashes <- c("00", "0000", "0001", "0010", "001001")
xyz <- geohashes_to_gridcells(geohashes)
geohashes == gridcells_to_geohashes(xyz)
}
\seealso{
\code{\link{gridcells_to_geohashes}} for the reverse conversion.

Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{gridcells_to_lnglat}}, \code{\link{itostr}},
  \code{\link{lnglat_to_gridcells}}
}
