% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{geohashes_to_latlng}
\alias{geohashes_to_latlng}
\title{Convert Geohashes to Lat, Lng}
\usage{
geohashes_to_latlng(geohashes)
}
\description{
Convert Geohashes to Lat, Lng
}
\examples{
geohashes_to_latlng(c("0000", "00"))
geohashes_to_latlng(latlng_to_geohashes(12.34, 153.21, 0))
geohashes_to_latlng(latlng_to_geohashes(12.34, 153.21, 13))
geohashes_to_latlng(latlng_to_geohashes(12.34, 153.21, 31))
}
\seealso{
Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{itostr}}, \code{\link{latlng_to_geohashes}},
  \code{\link{latlng_to_gridcells}}
}

