% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{geohashes_to_zoom}
\alias{geohashes_to_zoom}
\title{Calculate Zoom from Geohashes}
\usage{
geohashes_to_zoom(geohashes)
}
\arguments{
\item{geohashes}{Geohashes.}
}
\description{
Calculate Zoom from Geohashes
}
