% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{get_col_id}
\alias{get_col_id}
\title{Get Catalogue of Life (COL) ID}
\usage{
get_col_id(search_string, scientific_name = TRUE)
}
\description{
See documentation at http://webservice.catalogueoflife.org/col/webservice
}
\examples{
get_col_id("Malus domestica")
get_col_id("Abelmoschus")
get_col_id("Prunus")
}
\seealso{
Other COL functions: \code{\link{build_col_scientific_name}},
  \code{\link{get_col_page}},
  \code{\link{parse_col_common_names}},
  \code{\link{parse_col_scientific_names}}
}

