% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{get_col_page}
\alias{get_col_page}
\title{Get Catalogue of Life (COL) Page}
\usage{
get_col_page(id, content_only = TRUE)
}
\description{
See documentation at http://webservice.catalogueoflife.org/col/webservice.
}
\examples{
id <- get_col_id("Malus domestica")
str(get_col_page(id))
}
\seealso{
Other COL functions: \code{\link{build_col_scientific_name}},
  \code{\link{get_col_id}},
  \code{\link{parse_col_common_names}},
  \code{\link{parse_col_scientific_names}}
}

