% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{get_col_search}
\alias{get_col_search}
\title{Get Catalogue of Life (COL) Search Results}
\usage{
get_col_search(name, format = "json", response = "terse", start = 0)
}
\description{
See documentation at \url{http://webservice.catalogueoflife.org/col/webservice}.
}
\examples{
str(content(get_col_search("Malus domestica")))
str(content(get_col_search("Abelmoschus")))
str(content(get_col_search("Prunus")))
}
\seealso{
Other COL functions: \code{\link{build_col_scientific_name}},
  \code{\link{get_col_page}}, \code{\link{parse_col_page}},
  \code{\link{parse_col_search}}
}

