% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{get_eol_id}
\alias{get_eol_id}
\title{Get Encyclopedia of Life (EOL) ID}
\usage{
get_eol_id(search_string, exact = TRUE)
}
\description{
For reliable results, search string must be a scientific name.
See documentation at http://eol.org/api/docs/search.
}
\examples{
get_eol_id("Malus pumila")
get_eol_id("Abelmoschus")
get_eol_id("Prunus")
}
\seealso{
Other EOL functions: \code{\link{get_eol_page}},
  \code{\link{parse_eol_common_names}},
  \code{\link{parse_eol_scientific_names}}
}

