% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{get_eol_search}
\alias{get_eol_search}
\title{Get Encyclopedia of Life (EOL) Search Results}
\usage{
get_eol_search(q, page = 1, exact = TRUE, filter_by_taxon_concept_id,
  filter_by_hierarchy_entry_id, filter_by_string, cache_ttl)
}
\description{
See documentation at \url{http://eol.org/api/docs/search}. Alterations to default parameter values:
exact: FALSE => TRUE
}
\details{
NOTE: For reliable results, run an exact search (exact = TRUE) with a scientific name search string (q).
}
\examples{
str(content(get_eol_search("Malus domestica"))) # few equivalent results
str(content(get_eol_search("Abelmoschus"))) # one result
str(content(get_eol_search("Forbidden fruit"))) # no results
}
\seealso{
Other EOL functions: \code{\link{get_eol_page}},
  \code{\link{parse_eol_page}},
  \code{\link{parse_eol_search}}
}

