% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{get_inaturalist_id}
\alias{get_inaturalist_id}
\title{Get iNaturalist Taxon ID}
\usage{
get_inaturalist_id(search_string)
}
\description{
Get iNaturalist Taxon ID
}
\examples{
get_inaturalist_id("Malus domestica")
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_page}},
  \code{\link{parse_inaturalist_common_names}},
  \code{\link{parse_inaturalist_scientific_names}}
}

