% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{get_inaturalist_page}
\alias{get_inaturalist_page}
\title{Get iNaturalist Taxon Page}
\usage{
get_inaturalist_page(id, content_only = TRUE)
}
\description{
Get iNaturalist Taxon Page
}
\examples{
id <- get_inaturalist_id("Malus domestica")
str(get_inaturalist_page(id))
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_id}},
  \code{\link{parse_inaturalist_common_names}},
  \code{\link{parse_inaturalist_scientific_names}}
}

