% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{get_inaturalist_search}
\alias{get_inaturalist_search}
\title{Get iNaturalist Taxon Search Results}
\usage{
get_inaturalist_search(q, is_active = "true")
}
\arguments{
\item{is_active}{Either 'true', 'false', or 'any'.}
}
\description{
Undocumented, but equivalent to the HTML equivalent \url{http://www.inaturalist.org/taxa/search}.
}
\examples{
str(content(get_inaturalist_search("Malus domestica")))
str(content(get_inaturalist_search("Abelmoschus")))
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_page}},
  \code{\link{parse_inaturalist_page}},
  \code{\link{parse_inaturalist_search}}
}

