% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{get_page}
\alias{get_page}
\title{Get Any Page from URL}
\usage{
get_page(url, content_only = TRUE)
}
\description{
Get Any Page from URL
}
\examples{
str(get_page("https://en.wikipedia.org/wiki/Malus_domestica"))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_common_names}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikispecies_common_names}}
}

