% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{get_wiki_page}
\alias{get_wiki_page}
\title{Get Wiki Page}
\usage{
get_wiki_page(wiki, type, page)
}
\description{
Get Wiki Page
}
\examples{
str(get_wiki_page("en", "wikipedia", "Malus domestica"))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikipedia_names}},
  \code{\link{parse_wikispecies_common_names}}
}

