% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Google_cs_languages}
\alias{Google_cs_languages}
\title{Google Custom Search Languages}
\format{An object of class \code{character} of length 38.}
\source{
\url{https://developers.google.com/custom-search/docs/ref_languages}
}
\usage{
Google_cs_languages
}
\description{
List of the languages supported by Google Custom Search.
}
\seealso{
Other constants: \code{\link{Categories}},
  \code{\link{Earth_circumference}},
  \code{\link{Earth_radius}}, \code{\link{Language_codes}},
  \code{\link{Location_import_fields}},
  \code{\link{Taxonomic_ranks}}
}
\keyword{datasets}
