% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{gridcells_to_geohashes}
\alias{gridcells_to_geohashes}
\alias{geohashes_to_gridcells}
\title{Convert between Grid Cell Indices and Geohashes}
\usage{
gridcells_to_geohashes(xyz)

geohashes_to_gridcells(geohashes)
}
\arguments{
\item{xyz}{Rows of Web Mercator grid cell indices (x, y, zoom).}

\item{geohashes}{A character vector of geohashes.}
}
\description{
Converts between Web Mercator (\href{http://spatialreference.org/ref/sr-org/7483/}{EPSG:3857}) quadtree grid cell indices and binary geohashes.
}
\examples{
xi <- c(0, 0, 1, 0)
yi <- c(0, 1, 1, 0)
zoom <- c(0, 1, 1, 2)
xyz <- cbind(xi, yi, zoom)
geohashes <- gridcells_to_geohashes(xyz)
xyz <- geohashes_to_gridcells(geohashes)
geohashes == gridcells_to_geohashes(xyz)
}
