% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{gridcells_to_lnglat}
\alias{gridcells_to_lnglat}
\title{Convert Grid Cell Indices to Coordinates}
\usage{
gridcells_to_lnglat(xyz)
}
\arguments{
\item{xyz}{Rows of zero-based, integer grid cell indices (x, y, zoom).}
}
\value{
Rows of coordinates in decimal degrees (longitude, latitude) corresponding to the lower-left cell corners.
}
\description{
Converts Web Mercator (\href{http://spatialreference.org/ref/sr-org/7483/}{EPSG:3857}) quadtree grid cell indices to geographic WGS84 (\href{http://spatialreference.org/ref/epsg/wgs-84/}{EPSG:4326}) coordinates.
}
\examples{
xi <- c(0, 0, 1, 0)
yi <- c(0, 1, 1, 0)
zoom <- c(0, 1, 1, 2)
xyz <- cbind(xi, yi, zoom)
lnglat <- gridcells_to_lnglat(xyz)
xyz == lnglat_to_gridcells(lnglat, zoom)
}
\seealso{
\code{\link{lnglat_to_gridcells}} for the reverse conversion.

Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{geohashes_to_gridcells}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{itostr}}, \code{\link{lnglat_to_gridcells}}
}
