% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{itostr}
\alias{itostr}
\title{Convert Integers to Strings}
\usage{
itostr(x, base = 2, n_bits = 32)
}
\description{
Converts integers to strings according to the given base and the number of digits. Shorter strings are right-padded with zero. Reverse with base::strtoi().
}
\examples{
itostr(123, base = 2)
strtoi(itostr(123), base = 2)
}
\seealso{
Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{geohashes_to_latlng}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{latlng_to_geohashes}},
  \code{\link{latlng_to_gridcells}}
}

