% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{latlng_to_geohashes}
\alias{latlng_to_geohashes}
\title{Convert Grid Cell Index to Geohash}
\usage{
latlng_to_geohashes(lat, lng, grid_zoom)
}
\description{
Convert Grid Cell Index to Geohash
}
\examples{
latlng_to_geohashes(12.34, 153.21, 0)
latlng_to_geohashes(12.34, 153.21, 1)
latlng_to_geohashes(12.34, 153.21, 2)
latlng_to_geohashes(12.34, 153.21, 13)
}
\seealso{
Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{geohashes_to_latlng}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{itostr}}, \code{\link{latlng_to_gridcells}}
}

