% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{latlng_to_gridcells}
\alias{latlng_to_gridcells}
\title{Convert Lat,Lng to Grid Cell Indices}
\usage{
latlng_to_gridcells(lat, lng, grid_zoom)
}
\description{
Convert Lat,Lng to Grid Cell Indices
}
\examples{
latlng_to_gridcells(-45, -45, 1)
latlng_to_gridcells(45, -45, 1)
latlng_to_gridcells(45, 45, 1)
latlng_to_gridcells(-45, 45, 1)
}
\seealso{
Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{geohashes_to_latlng}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{itostr}}, \code{\link{latlng_to_geohashes}}
}

