% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{lnglat_to_gridcells}
\alias{lnglat_to_gridcells}
\title{Convert Coordinates to Grid Cell Indices}
\usage{
lnglat_to_gridcells(lnglat, zoom = 0L)
}
\arguments{
\item{lnglat}{Rows of coordinates in decimal degrees (longitude, latitude), with longitude: [-180, 180) and latitude: [-85.05113, 85.05113).}

\item{zoom}{Zoom level(s) of the quadtree dividing the Earth into a square \eqn{2^{zoom}}{2^zoom} grid.}
}
\value{
Rows of zero-based, integer grid cell indices (x, y, zoom).
}
\description{
Converts geographic WGS84 (\href{http://spatialreference.org/ref/epsg/wgs-84/}{EPSG:4326}) coordinates to Web Mercator (\href{http://spatialreference.org/ref/sr-org/7483/}{EPSG:3857}) quadtree grid cell indices.
}
\examples{
lng <- c(-45, -45, 45, 45)
lat <- c(-45, 45, 45, -45)
lnglat <- cbind(lng, lat)
lnglat_to_gridcells(lnglat, zoom = 1)
}
\seealso{
Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{geohashes_to_gridcells}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{gridcells_to_lnglat}}, \code{\link{itostr}}
}
