% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{lnglat_to_xy}
\alias{lnglat_to_xy}
\alias{xy_to_lnglat}
\title{Convert between WGS84 and Web Mercator Coordinates}
\usage{
lnglat_to_xy(lnglat)

xy_to_lnglat(xy)
}
\arguments{
\item{lnglat}{Rows of coordinates in decimal degrees (longitude, latitude), with longitude: [-180, 180) and latitude: [-85.05112, 85.05112].}

\item{xy}{Rows of Web Mercator coordinates in meters (x, y).}
}
\description{
Converts between geographic WGS84 (\href{http://spatialreference.org/ref/epsg/wgs-84/}{EPSG:4326}) coordinates in decimal degrees and projected Web Mercator (\href{http://spatialreference.org/ref/sr-org/7483/}{EPSG:3857}) coordinates in meters.
}
\examples{
lng <- c(-45, -45, 45, 45)
lat <- c(-45, 45, 45, -45)
lnglat <- cbind(lng, lat)
xy_to_lnglat(lnglat_to_xy(lnglat)) - lnglat
}
