% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Location_import_fields}
\alias{Location_import_fields}
\title{Falling Fruit Location import fields}
\format{An object of class \code{character} of length 15.}
\usage{
Location_import_fields
}
\description{
Field names expected by Falling Fruit for locations import.
}
\seealso{
Other constants: \code{\link{Categories}},
  \code{\link{Earth_circumference}},
  \code{\link{Earth_radius}},
  \code{\link{Google_cs_languages}},
  \code{\link{Language_codes}},
  \code{\link{Taxonomic_ranks}}
}
\keyword{datasets}
