% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lvapply}
\alias{lvapply}
\title{Apply a Function over a List of Atomic Vectors}
\usage{
lvapply(X, FUN, ...)
}
\arguments{
\item{X}{List of atomic vectors.}

\item{FUN}{Function to be applied to each element of \code{X}.}

\item{...}{Arguments passed to \code{FUN}.}
}
\value{
Object of the same length as \code{X}.
}
\description{
Provides a fast, vectorized alternative to \code{\link{lapply}} for a list of atomic vectors operated on by a function whose result is the same length as its input.
}
\examples{
x <- list(c("a", "b"), c(), c("d"))
lvapply(x, gsub, pattern = "d", replacement = "e")
}
\seealso{
Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{melt_by_listcol}},
  \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sptransform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
