% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{match_names_to_ff_types}
\alias{match_names_to_ff_types}
\title{Match Names Against Falling Fruit Types}
\usage{
match_names_to_ff_types(scientific_names = NULL, common_names = NULL,
  types = get_ff_types(pending = FALSE), max_distance = 3,
  n_nearest = Inf, locale = "en", ...)
}
\arguments{
\item{scientific_names}{Vector of scientific names.}

\item{common_names}{Vector of common names.}

\item{types}{Falling Fruit types, as returned by \code{\link{get_ff_types}}.}

\item{max_distance}{Maximum distance for fuzzy matching, as returned by \code{\link{stringdist::stringdistmatrix}}.}

\item{locale}{Locale of names to use for matching.}

\item{...}{Additional parameters passed to \code{\link{stringdist::stringdistmatrix}}.}
}
\description{
Match Names Against Falling Fruit Types
}
\seealso{
Other location import functions: \code{\link{aggregate_locations_by_position}},
  \code{\link{apply_ff_type_matches}},
  \code{\link{build_match_table}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}

