% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{match_to_ff_types}
\alias{match_to_ff_types}
\title{Match Against Falling Fruit Types}
\usage{
match_to_ff_types(dt, types, saved_match_table)
}
\description{
TODO: Don't match if already matched.
}
\seealso{
Other location import functions: \code{\link{apply_ff_type_matches}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}

