% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{match_type_strings}
\alias{match_type_strings}
\title{Match Type Strings to Types}
\usage{
match_type_strings(type_strings, types, simplify = FALSE)
}
\description{
Match Type Strings to Types
}
\examples{
types <- get_ff_types()
match_type_strings("Apple", types)
match_type_strings(c("Apple [Malus domestica]", "Pear [Pyrus]"), types)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}

