% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{merge_overlapping_locations}
\alias{merge_overlapping_locations}
\title{Aggregate Locations by Position}
\usage{
merge_overlapping_locations(dt, ...)
}
\arguments{
\item{dt}{Locations data.}

\item{...}{Arguments passed to \code{\link{build_location_descriptions}}.}
}
\description{
Falling Fruit does not support overlapping locations. This function merges locations with the same \code{lat}, \code{lng} or \code{address} by the methods described in Details.
}
\details{
The following aggregation is applied to all non-missing values for each field:

\itemize{
  \item \code{ids, types, author} - Comma-delimited list of all unique values.
  \item \code{description} - Result of \code{\link{build_location_descriptions}(types = description, notes = notes)}.
  \item \code{season.start} - Minimum value.
  \item \code{season.stop} - Maximum value.
  \item \code{unverified} - True ('x') if any are true.
  \item \code{yield.rating, quality.rating} - Rounded mean of all values.
  \item \code{photo.url} - First value.
}
}
\seealso{
Other location import functions: \code{\link{apply_match_table}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}
