% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{na.remove}
\alias{na.remove}
\title{Remove Missing Values from Object}
\usage{
na.remove(x)
}
\arguments{
\item{x}{Object passed to \code{\link[stats]{na.omit}}.}
}
\value{
\code{x} with \code{\link[base]{NA}}s removed.
}
\description{
Silently removes missing values or incomplete cases with \code{\link[base]{NA}}s.
}
\examples{
str(na.omit(c(1, NA)))
str(na.remove(c(1, NA)))
}
\seealso{
Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{lvapply}},
  \code{\link{melt_by_listcol}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sptransform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
