% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{normalize_common_name}
\alias{normalize_common_name}
\title{Normalize common names}
\usage{
normalize_common_name(x, x_lower = NULL, x_search = NULL)
}
\arguments{
\item{x}{(character vector) Common names to normalize.}

\item{x_lower}{(character vector) Lower case \code{x} (precomputed for speed).}

\item{x_search}{(character vector) Search-formatted \code{x} (precomputed for speed).}
}
\description{
Normalize common names
}
\seealso{
Other translation functions: \code{\link{build_common_name_table}},
  \code{\link{build_scientific_name_table}},
  \code{\link{parse_sources_about_type}},
  \code{\link{query_sources_about_type}}
}
