% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_language}
\alias{normalize_language}
\title{Normalize Language}
\usage{
normalize_language(x, type = "code")
}
\description{
Returns the highest level ISO language code corresponding to the ISO language code, language name, autonym, or Wikipedia provided. If the input matches zero or multiple entries, the input is returned and a warning is raised.
}
\examples{
normalize_language("spa", "code")
normalize_language("spanish", "en")
normalize_language("español", "autonym")
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{replace_values_in_list}},
  \code{\link{unique_na}}
}

