% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{parse_col_page}
\alias{parse_col_page}
\title{Parse Catalogue of Life (COL) Page}
\usage{
parse_col_page(page, types = c("scientific_names", "common_names"))
}
\description{
Parse Catalogue of Life (COL) Page
}
\examples{
s <- get_col_search("Malus domestica")
id <- parse_col_search(s, "ids")$ids[1]
pg <- get_col_page(id)
str(parse_col_page(pg))
}
\seealso{
Other COL functions: \code{\link{build_col_scientific_name}},
  \code{\link{get_col_page}}, \code{\link{get_col_search}},
  \code{\link{parse_col_search}}
}

