% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{parse_eol_common_names}
\alias{parse_eol_common_names}
\title{Parse Encyclopedia of Life (EOL) Common Names}
\usage{
parse_eol_common_names(json)
}
\description{
Parse Encyclopedia of Life (EOL) Common Names
}
\examples{
id <- get_eol_id("Malus domestica")
json <- get_eol_page(id)
parse_eol_common_names(json)
}
\seealso{
Other EOL functions: \code{\link{get_eol_id}},
  \code{\link{get_eol_page}},
  \code{\link{parse_eol_scientific_names}}
}

