% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{parse_eol_page}
\alias{parse_eol_page}
\title{Parse Encyclopedia of Life (EOL) Page}
\usage{
parse_eol_page(page, types = c("scientific_names", "common_names"))
}
\description{
Parse Encyclopedia of Life (EOL) Page
}
\examples{
s <- get_eol_search("Malus domestica")
id <- parse_eol_search(s, "ids")$ids[1]
pg <- get_eol_page(id)
str(parse_eol_page(pg))
}
\seealso{
Other EOL functions: \code{\link{get_eol_page}},
  \code{\link{get_eol_search}},
  \code{\link{parse_eol_search}}
}

