% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{parse_eol_search}
\alias{parse_eol_search}
\title{Parse Encyclopedia of Life (EOL) Search Results}
\usage{
parse_eol_search(search, types = c("results", "ids"))
}
\description{
Parse Encyclopedia of Life (EOL) Search Results
}
\examples{
s <- get_eol_search("Malus domestica")
str(parse_eol_search(s)) # few equivalent results
s <- get_eol_search("Abelmoschus")
str(parse_eol_search(s)) # one result
s <- get_eol_search("Forbidden fruit")
str(parse_eol_search(s)) # no results
}
\seealso{
Other EOL functions: \code{\link{get_eol_page}},
  \code{\link{get_eol_search}},
  \code{\link{parse_eol_page}}
}

