% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{parse_inaturalist_page}
\alias{parse_inaturalist_page}
\title{Parse iNaturalist Taxon Page}
\usage{
parse_inaturalist_page(page, types = c("scientific_names", "common_names"))
}
\description{
Parse iNaturalist Taxon Page
}
\examples{
s <- get_inaturalist_search("Malus domestica")
id <- parse_inaturalist_search(s, "ids")$ids[1]
pg <- get_inaturalist_page(id)
str(parse_inaturalist_page(pg))
s <- get_inaturalist_search("Abelmoschus")
id <- parse_inaturalist_search(s, "ids")$ids[1]
pg <- get_inaturalist_page(id)
str(parse_inaturalist_page(pg))
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_page}},
  \code{\link{get_inaturalist_search}},
  \code{\link{parse_inaturalist_search}}
}

