% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{parse_inaturalist_scientific_names}
\alias{parse_inaturalist_scientific_names}
\title{Parse iNaturalist Scientific Names}
\usage{
parse_inaturalist_scientific_names(json)
}
\description{
Parse iNaturalist Scientific Names
}
\examples{
id <- get_inaturalist_id("Malus domestica")
json <- get_inaturalist_page(id)
parse_inaturalist_scientific_names(json)
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_id}},
  \code{\link{get_inaturalist_page}},
  \code{\link{parse_inaturalist_common_names}}
}

