% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{parse_inaturalist_search}
\alias{parse_inaturalist_search}
\title{Parse iNaturalist Taxon Search Results}
\usage{
parse_inaturalist_search(search, types = c("results", "ids"), exact = TRUE,
  scientific_name = TRUE, ignore.case = TRUE)
}
\description{
Parse iNaturalist Taxon Search Results
}
\examples{
s <- get_inaturalist_search("Malus domestica")
str(parse_inaturalist_search(s))
s <- get_inaturalist_search("Abelmoschus")
str(parse_inaturalist_search(s))
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_page}},
  \code{\link{get_inaturalist_search}},
  \code{\link{parse_inaturalist_page}}
}

