% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_scientific_names}
\alias{parse_scientific_names}
\title{Parse Scientific Names}
\usage{
parse_scientific_names(x, parts = c("taxon", "subtaxons", "connectors",
  "cultivars"))
}
\arguments{
\item{x}{Character vector.}

\item{parts}{Parts of the scientific names to extract.}
}
\value{
List of character vectors (for "taxon") and lists of the same length as \code{x}.
}
\description{
Extracts components of scientific names.
}
\examples{
parse_scientific_names(c("Malus pumila", "Prunus persica"), parts = "taxon")
parse_scientific_names("Prunus persica var. nucipersica 'Crimson King'")
}
\seealso{
Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{lvapply}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sp_transform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
