% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{parse_sources_about_type}
\alias{parse_sources_about_type}
\title{Parse sources about Falling Fruit type}
\usage{
parse_sources_about_type(responses)
}
\arguments{
\item{responses}{(response list) Result of \code{\link{query_sources_about_type}}.}
}
\description{
Parse sources about Falling Fruit type
}
\seealso{
Other translation functions: \code{\link{build_common_name_table}},
  \code{\link{build_scientific_name_table}},
  \code{\link{normalize_common_name}},
  \code{\link{query_sources_about_type}}
}
