% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{parse_type_strings}
\alias{parse_type_strings}
\title{Parse Type Strings}
\usage{
parse_type_strings(type_strings)
}
\arguments{
\item{type_strings}{Character vector of type strings.}
}
\value{
List of lists with id, name, and scientific name.
}
\description{
Parses type strings into their component parts.
}
\examples{
parse_type_strings("14")
parse_type_strings("Apple [Malus pumila]")
parse_type_strings("14: Apple [Malus pumila]")
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_descriptions}},
  \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}}
}
