% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wiki_url}
\alias{parse_wiki_url}
\title{Parse Wiki URL}
\usage{
parse_wiki_url(url)
}
\description{
Parse Wiki URL
}
\examples{
parse_wiki_url("https://en.wikipedia.org/wiki/Malus_domestica")
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikipedia_names}},
  \code{\link{parse_wikispecies_common_names}}
}

