% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikicommons_common_names}
\alias{parse_wikicommons_common_names}
\title{Parse Wikimedia Commons Common Names}
\usage{
parse_wikicommons_common_names(xml)
}
\description{
Parse Wikimedia Commons Common Names
}
\examples{
xml <- content(GET("https://commons.wikimedia.org/wiki/Malus_domestica"))
parse_wikicommons_common_names(xml)
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikipedia_names}},
  \code{\link{parse_wikispecies_common_names}}
}

