% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikipedia_common_names}
\alias{parse_wikipedia_common_names}
\title{Parse Wikipedia Common Names}
\usage{
parse_wikipedia_common_names(xml,
  language = parse_wiki_url(xml_attr(xml_find_first(xml, xpath =
  "//link[@rel='canonical']"), "href"))[[1]])
}
\description{
Parse Wikipedia Common Names
}
\examples{
xml <- get_page("https://en.wikipedia.org/wiki/Malus_domestica")
parse_wikipedia_common_names(xml)
xml <- get_page("https://en.wikipedia.org/wiki/Abelmoschus")
parse_wikipedia_common_names(xml) # no names
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_page}}, \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikispecies_common_names}}
}

