% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikipedia_langlinks}
\alias{parse_wikipedia_langlinks}
\title{Parse Wikipedia Language Links}
\usage{
parse_wikipedia_langlinks(json)
}
\description{
Parse Wikipedia Language Links
}
\examples{
json <- get_wiki_page("en", "wikipedia", "Malus domestica")
parse_wikipedia_langlinks(json)
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_names}},
  \code{\link{parse_wikispecies_common_names}}
}

