% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikipedia_names}
\alias{parse_wikipedia_names}
\title{Parse Wikipedia Names}
\usage{
parse_wikipedia_names(xml)
}
\description{
Parse Wikipedia Names
}
\examples{
xml <- content(GET("https://en.wikipedia.org/wiki/Malus_domestica"))
parse_wikipedia_names(xml)
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_common_names}},
  \code{\link{parse_wikipedia_langlinks}},
  \code{\link{parse_wikispecies_common_names}}
}

