% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikipedia_page}
\alias{parse_wikipedia_page}
\title{Parse Wikipedia Page}
\usage{
parse_wikipedia_page(page, types = c("common_names", "langlinks"))
}
\arguments{
\item{page}{(response) Result of \code{\link{get_wiki_page}}.}

\item{types}{(character) Data types to parse.}
}
\description{
Parse Wikipedia Page
}
\examples{
pg <- get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")
str(parse_wikipedia_page(pg))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikispecies_page}}
}
