% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikispecies_page}
\alias{parse_wikispecies_page}
\title{Parse Wikispecies Page}
\usage{
parse_wikispecies_page(page, types = c("common_names"))
}
\description{
Parse Wikispecies Page
}
\examples{
pg <- get_wiki_page("https://species.wikimedia.org/wiki/Malus_domestica")
str(parse_wikispecies_page(pg))
pg <- get_wiki_page("https://species.wikimedia.org/wiki/Abelmoschus")
str(parse_wikispecies_page(pg)) # no names
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikipedia_page}}
}

