% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{read_locations}
\alias{read_locations}
\title{Load Locations Dataset}
\usage{
read_locations(file, latlng = c("lat", "lng"), id = "id",
  CRSobj = CRS("+proj=longlat +ellps=WGS84"), ...)
}
\description{
TODO: Check that id field is unique, warn otherwise.
}
\seealso{
Other location import functions: \code{\link{apply_ff_type_matches}},
  \code{\link{match_to_ff_types}},
  \code{\link{write_locations_for_import}}
}

