% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_values_in_list}
\alias{replace_values_in_list}
\title{Replace Values in List}
\usage{
replace_values_in_list(x, old, new)
}
\description{
Replace Values in List
}
\examples{
x <- list(NULL, list(NULL, NA))
str(replace_values_in_list(x, NULL, 1))
str(replace_values_in_list(x, list(NULL, NA), 1))
str(replace_values_in_list(x, list(NULL, NA), list(1, 2)))
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}}, \code{\link{unique_na}}
}

