% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sptransform}
\alias{sptransform}
\title{Transform Spatial Coordinates}
\usage{
sptransform(xy, from = tryCatch(sp::proj4string(xy)), to = 4326)
}
\arguments{
\item{xy}{Rows of coordinates (x, y).}

\item{from}{Initial projection as numeric (EPSG code), character (proj4 string), or \code{\link[sp]{CRS}}.}

\item{to}{Target projection as numeric (EPSG code), character (proj4 string), or \code{\link[sp]{CRS}}.}
}
\value{
Transformed coordinates coerced to the same class as \code{xy}.
}
\description{
Transform Spatial Coordinates
}
\examples{
from <- 21781
to <- 4326
sptransform(sptransform(c(500, 1000), from, to), to, from)
}
\seealso{
\code{\link[sp]{spTransform}}

Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{lvapply}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
