% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{subset_search_results}
\alias{subset_search_results}
\title{Subset Search Results}
\usage{
subset_search_results(strings, values, ignore.case = TRUE)
}
\description{
Difference search results based on subsetting of search strings (interpreted as search phrases).
}
\examples{
# Pine: 20, Blue pine: 17
strings <- c("Pine", "Blue pine")
results <- c(20, 17)
subset_search_results(strings, results) # 3, 17
# Pine: 20, Blue pine: 17, White blue pine: 10, White: 20
strings <- c("Pine", "Blue pine", "White blue pine", "White")
results <- c(20, 17, 10, 20)
subset_search_results(strings, results) # 3, 7, 10, 10
# Pine: 20, Blue pine: 17, Blue pine a: 10, Blue pine b: 5, Blue pine b c: 1
strings <- c("Pine", "Blue pine", "Blue pine a", "Blue pine b", "Blue pine b c")
results <- c(20, 17, 10, 5, 1)
subset_search_results(strings, results) # 3, 2, 10, 4, 1
# Pine: 20, Blue: 15, Blue pine: 10, Pine blue: 3
strings <- c("Pine", "Blue", "Blue pine", "Pine blue")
results <- c(20, 15, 10, 3)
subset_search_results(strings, results) # 7, 2, 10, 3
# Pine: 200, Blue pine: 100, White pine: 100
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{unique_na}}
}

