% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Taxonomic_ranks}
\alias{Taxonomic_ranks}
\title{Falling Fruit Taxonomic ranks}
\format{An object of class \code{character} of length 10.}
\usage{
Taxonomic_ranks
}
\description{
Names of the taxonomic ranks returned by Falling Fruit as zero-based indices.
}
\seealso{
Other constants: \code{\link{Categories}},
  \code{\link{Earth_circumference}},
  \code{\link{Earth_radius}},
  \code{\link{Google_cs_languages}},
  \code{\link{Language_codes}},
  \code{\link{Location_import_fields}}
}
\keyword{datasets}
