% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Unicode}
\alias{Unicode}
\title{Unicode characters by locale}
\format{A named list of character vectors.}
\description{
\itemize{
  \item zh_hant: Traditional chinese characters (\url{https://github.com/jpatokal/script_detector/blob/master/lib/chinese_detector.rb})
  \item zh_hans: Simplified chinese characters (\url{https://github.com/jpatokal/script_detector/blob/master/lib/chinese_detector.rb})
}
}
\seealso{
Other constants: \code{\link{Categories}},
  \code{\link{Earth_circumference}},
  \code{\link{Earth_radius}},
  \code{\link{Google_cs_languages}},
  \code{\link{Language_codes}},
  \code{\link{Location_import_fields}},
  \code{\link{Taxonomic_ranks}}
}
