% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unique_na}
\alias{unique_na}
\title{Return Single Unique Value or NA}
\usage{
unique_na(x, na.rm = FALSE)
}
\description{
Return Single Unique Value or NA
}
\examples{
unique_na(c(1, 1))
unique_na(c(1, 2))
unique_na(c(1, NA), na.rm = FALSE)
unique_na(c(1, NA), na.rm = TRUE)
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}}
}

