% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{xy_to_gridcells}
\alias{xy_to_gridcells}
\alias{gridcells_to_xy}
\title{Convert between Web Mercator Coordinates and Grid Cell Indices}
\usage{
xy_to_gridcells(xy, zoom = 0L)

gridcells_to_xy(xyz)
}
\arguments{
\item{xy}{Rows of Web Mercator coordinates in meters (x, y).}

\item{zoom}{Zoom level(s) of the quadtree dividing the Earth into a square \eqn{2^{zoom}}{2^zoom} grid.}

\item{xyz}{Rows of zero-based, integer grid cell indices (x, y, zoom).}
}
\description{
Converts between projected Web Mercator (\href{http://spatialreference.org/ref/sr-org/7483/}{EPSG:3857}) coordinates in meters and quadtree grid cell indices. The conversion back from grid cell indices returns the coordinates of the lower left corner of the cell.
}
\examples{
xy <- cbind(x = 0, y = 0)
xyz <- xy_to_gridcells(xy, zoom = 1)
gridcells_to_xy(xyz)
xyz <- xy_to_gridcells(xy, zoom = 0)
gridcells_to_xy(xyz)
}
