% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{build_col_scientific_name}
\alias{build_col_scientific_name}
\title{Build Catalogue of Life (COL) Scientific Name}
\usage{
build_col_scientific_name(json)
}
\description{
Build Catalogue of Life (COL) Scientific Name
}
\examples{
id <- get_col_id("Malus domestica")
json <- get_col_page(id)
build_col_scientific_name(json)
str(lapply(json$synonyms, build_col_scientific_name))
}
\seealso{
Other COL functions: \code{\link{get_col_id}},
  \code{\link{get_col_page}},
  \code{\link{parse_col_common_names}},
  \code{\link{parse_col_scientific_names}}
}

