% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{build_location_description}
\alias{build_location_description}
\title{Build Location Description}
\usage{
build_location_description(type_strings, notes = NULL, merge = FALSE,
  type_sep = ", ", note_sep = ". ", group_sep = "<br>",
  frequency = TRUE, frequency_in = "[]")
}
\arguments{
\item{type_strings}{Character vector interpreted literally (comma-delimited lists are not split).}

\item{notes}{List of character vectors, each of the same length.}

\item{merge}{Whether to merge types, discarding any notes that are not equal for all types.}

\item{type_sep}{Character string to seperate each element in \code{type_strings} when \code{merge = TRUE}.}

\item{note_sep}{Character string to seperate each retained element in \code{notes}.}

\item{group_sep}{Character string to seperate each grouping when \code{merge = FALSE}.}

\item{frequency}{Whether to display frequency before each element in \code{type_strings}.}

\item{frequency_in}{String of two characters in which to display the frequencies.}
}
\value{
Character string of the summarized and concatenated values.
}
\description{
Builds a description from its inputs. When \code{merge} is \code{TRUE}, all content is summarized as one group, and all notes not equal throughout are discarded. Otherwise, all content is preserved and only grouped for equal types with all equal notes.
}
\examples{
type_strings <- c("Apple", "Pear", "Pear")
notes <- list(
  c("Planted 1999", "Height 10 m"),
  c("Planted 1999", "Height 20 m"),
  c("Planted 1999", "Height 20 m")
)
build_location_description(type_strings, notes)
build_location_description(type_strings, notes, merge = TRUE)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_descriptions}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}
