% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{capitalize_words}
\alias{capitalize_words}
\title{Capitalize Words
Skips numbers at start of words (e.g. 3-in-1 pear).}
\usage{
capitalize_words(x, strict = FALSE, first = FALSE)
}
\description{
Capitalize Words
Skips numbers at start of words (e.g. 3-in-1 pear).
}

