% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_strings}
\alias{clean_strings}
\title{Clean Strings}
\usage{
clean_strings(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector of the same length as \code{x}.
}
\description{
Performs generic string cleaning operations.
}
\examples{
clean_strings("[](){}")
clean_strings("``''\\"\\"")
clean_strings("..,,")
clean_strings(" Hello,,  `world`.. how are you?()")
}
\seealso{
Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{lvapply}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sptransform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
