% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{count_google_cs_results}
\alias{count_google_cs_results}
\title{Count Google Custom Search (CS) Results}
\source{
\url{https://developers.google.com/custom-search/json-api/v1/overview}
}
\usage{
count_google_cs_results(string, language = NULL, pause = FALSE)
}
\arguments{
\item{string}{(character) Search string.}

\item{language}{(character) Language code to search in.}

\item{pause}{(boolean) Whether to pause 1 s.}
}
\description{
Free up to 100/day. Maximum 10,000/day. Rate limit 1/s.
}
\examples{
count_google_cs_results("'Malus domestica'+'Apfel'", "en")
count_google_cs_results("'Malus domestica'+'Apfel'", "de")
}
