% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{expand_category_mask}
\alias{expand_category_mask}
\title{Expand Binary Category Mask to Categories}
\usage{
expand_category_mask(category_mask)
}
\description{
Expand Binary Category Mask to Categories
}
\examples{
expand_category_mask(0)
expand_category_mask(3)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}

