% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{expand_geohashes}
\alias{expand_geohashes}
\title{Expand Geohashes}
\usage{
expand_geohashes(geohashes)
}
\description{
Expand Geohashes
}
\examples{
expand_geohashes("0011")
expand_geohashes(c("0011", "110011"))
geohashes_to_latlng(expand_geohashes(latlng_to_geohashes(12.34, 153.21, 31)))
}
\seealso{
Other cluster functions: \code{\link{geohashes_to_latlng}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{itostr}}, \code{\link{latlng_to_geohashes}},
  \code{\link{latlng_to_gridcells}}
}

