% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{expand_geohashes}
\alias{expand_geohashes}
\title{Expand Geohashes}
\usage{
expand_geohashes(geohashes)
}
\arguments{
\item{geohashes}{A character vector of geohashes.}
}
\value{
Character vector of each geohash in \code{geohashes} followed by its parents.
}
\description{
Returns geohashes followed by all their parent geohashes.
}
\examples{
expand_geohashes("0011")
expand_geohashes(c("0011", "110011"))
xyz <- lnglat_to_gridcells(c(123.4567, 12.3456), 31)
geohashes <- gridcells_to_geohashes(xyz)
exyz <- geohashes_to_gridcells(expand_geohashes(geohashes))
lnglat <- gridcells_to_lnglat(exyz)
cbind(exyz, lnglat)
}
\seealso{
Other cluster functions: \code{\link{geohashes_to_gridcells}},
  \code{\link{gridcells_to_geohashes}},
  \code{\link{gridcells_to_lnglat}}, \code{\link{itostr}},
  \code{\link{lnglat_to_gridcells}}
}
