% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_addresses}
\alias{format_addresses}
\title{Format Addresses}
\usage{
format_addresses(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Ideally, correctly formatted addresses.
}
\description{
Attempts to clean and format addresses names by the following convention:\cr
<Number> <SE|SW|NE|NW> <Street McName> <Ave|*>
}
\examples{
format_addresses("123 SE. MCDONALD AV")
}
\seealso{
\code{\link{clean_strings}} for generic string cleaning.

Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{melt_by_listcol}},
  \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sptransform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
