% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_scientific_names}
\alias{format_scientific_names}
\title{Format Specialty Strings}
\usage{
format_scientific_names(x, connecting_terms = TRUE, cultivars = TRUE)
}
\description{
Supports species common names, latin names, and addresses.
}
\examples{
format_scientific_names("Malus")
format_scientific_names("Malus Batt.")
format_scientific_names("Tamarix rubella Batt.")
format_scientific_names("Tamarix lucronensis Sennen & Elias")
format_scientific_names("Malus ×domestica var. gala", connecting_terms = TRUE)
format_scientific_names("Malus ×domestica var. gala", connecting_terms = FALSE)
format_scientific_names("Malus pumila 'gala'", cultivars = TRUE)
format_scientific_names("Malus pumila 'gala'", cultivars = FALSE)
format_scientific_names("Prunus subg amygdalus")
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_addresses}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{sptransform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
