% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{gct_translate}
\alias{gct_translate}
\title{Translate text}
\source{
\url{https://cloud.google.com/translate/docs/translating-text}
}
\usage{
gct_translate(q, source = NULL, target, format = "html", key)
}
\arguments{
\item{q}{(character vector) Text strings to translate.}

\item{source}{(character) Source language code. If \code{NULL}, it is detected automatically.}

\item{target}{(character) Target language code.}

\item{format}{(character) Format of \code{q}: "html" or "text".}

\item{key}{(character) API key}
}
\description{
Translates text strings using the Google Cloud Translation API.
}
