% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{get_eol_id}
\alias{get_eol_id}
\title{Get Encyclopedia of Life (EOL) ID}
\usage{
get_eol_id(search_string)
}
\description{
Get Encyclopedia of Life (EOL) ID
}
\examples{
get_eol_id("Malus domestica")
}
\seealso{
Other EOL functions: \code{\link{get_eol_page}},
  \code{\link{parse_eol_common_names}},
  \code{\link{parse_eol_scientific_names}}
}

