% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eol.R
\name{get_eol_page}
\alias{get_eol_page}
\title{Get Encyclopedia of Life (EOL) Page}
\usage{
get_eol_page(id, common_names = TRUE, synonyms = TRUE, references = TRUE,
  taxonomy = TRUE, details = TRUE, iucn = TRUE, images = 75,
  content_only = TRUE)
}
\description{
See documentation at http://eol.org/api/docs/pages.
}
\examples{
id <- get_eol_id("Malus domestica")
str(get_eol_page(id))
}
\seealso{
Other EOL functions: \code{\link{get_eol_id}},
  \code{\link{parse_eol_common_names}},
  \code{\link{parse_eol_scientific_names}}
}

