% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{get_ff_types}
\alias{get_ff_types}
\title{Get Falling Fruit (FF) Types}
\usage{
get_ff_types(categories = c("forager", "freegan", "honeybee", "grafter"),
  uncategorized = TRUE, pending = TRUE, urls = TRUE, locale = "en")
}
\description{
Get Falling Fruit (FF) Types
}
\examples{
types <- get_ff_types(pending = FALSE)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}},
  \code{\link{parse_type_strings}}
}

