% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaturalist.R
\name{get_inaturalist_page}
\alias{get_inaturalist_page}
\title{Get iNaturalist Taxon Page}
\usage{
get_inaturalist_page(id)
}
\arguments{
\item{id}{iNaturalist page ID.}
}
\description{
Undocumented, but equivalent to the HTML equivalent \url{http://www.inaturalist.org/taxa/:id.json}.
}
\examples{
s <- get_inaturalist_search("Malus domestica")
id <- parse_inaturalist_search(s, "ids")$ids[1]
pg <- get_inaturalist_page(id)
str(httr::content(pg))
}
\seealso{
Other iNaturalist functions: \code{\link{get_inaturalist_search}},
  \code{\link{parse_inaturalist_page}},
  \code{\link{parse_inaturalist_search}}
}
