% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{get_wiki_page}
\alias{get_wiki_page}
\title{Get Wiki Page from API}
\usage{
get_wiki_page(url, wiki, type, page, format = "json", action = "parse",
  redirects = TRUE)
}
\description{
Get Wiki Page from API
}
\examples{
str(content(get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikipedia_page}},
  \code{\link{parse_wikispecies_page}}
}

