% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{get_wiki_page}
\alias{get_wiki_page}
\title{Get Wiki Page from API}
\usage{
get_wiki_page(url, format = "json", action = "parse", redirects = TRUE)
}
\arguments{
\item{url}{(character) Wiki page URL.}

\item{format}{(character) Response format.}

\item{action}{(character) Response action.}

\item{redirects}{(boolean) Whether to follow redirects.}
}
\description{
Get Wiki Page from API
}
\examples{
pg <- get_wiki_page("https://en.wikipedia.org/wiki/Malus_domestica")
str(httr::content(pg))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikicommons_page}},
  \code{\link{parse_wikipedia_page}},
  \code{\link{parse_wikispecies_page}}
}
