% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{gridcells_to_geohashes}
\alias{gridcells_to_geohashes}
\title{Convert Grid Cell Indices to Geohash}
\usage{
gridcells_to_geohashes(grid_cells)
}
\description{
Convert Grid Cell Indices to Geohash
}
\examples{
gridcells_to_geohashes(cbind(0, 1, 1))
}
\seealso{
Other cluster functions: \code{\link{expand_geohashes}},
  \code{\link{geohashes_to_latlng}}, \code{\link{itostr}},
  \code{\link{latlng_to_geohashes}},
  \code{\link{latlng_to_gridcells}}
}

