% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.empty}
\alias{is.empty}
\title{Check if Object is Empty}
\usage{
is.empty(x, env = globalenv())
}
\description{
FIXME: Returns FALSE if object doesn't exist when called within other function.
}
\examples{
is.empty(list())
is.empty(list(NULL, NA, "", c()))
is.empty(list(1, NULL, NA, "", c()))
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{replace_values_in_list}},
  \code{\link{unique_na}}
}

