% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{merge_overlapping_locations}
\alias{merge_overlapping_locations}
\title{Aggregate Locations by Position}
\usage{
merge_overlapping_locations(dt, frequency = TRUE, note_sep = ". ")
}
\arguments{
\item{dt}{Locations data.}

\item{frequency}{Whether to include the number of each types in the description.}

\item{sep}{Character string to seperate the retained notes.}
}
\description{
NOTE: Multi-type locations may need to be split before being merged (see \code{\link{build_location_description}}).
}
\seealso{
Other location import functions: \code{\link{apply_match_table}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{read_locations}},
  \code{\link{write_locations_for_import}}
}
