% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_language}
\alias{normalize_language}
\title{Normalize Language}
\usage{
normalize_language(x, types = c("ISO639.1", "ISO639.2B", "ISO639.2T",
  "ISO639.3", "ISO639.6", "ISO639.3_macro", "wikipedia", "autonym", "en", "fr",
  "de", "ru", "es", "zh"))
}
\description{
Returns the highest level ISO or Wikipedia language code corresponding to the ISO or Wikipedia language code, language name, or autonym provided. If the input matches zero or multiple entries, the input is returned and a warning is raised.
}
\examples{
normalize_language("spa")
normalize_language("Spanish")
normalize_language("Español")
normalize_language("Espagnol")
}
\seealso{
Other helper functions: \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{expand_list_by_split}},
  \code{\link{format_strings}}, \code{\link{is.empty}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{replace_values_in_list}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}

