% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{normalize_type_strings}
\alias{normalize_type_strings}
\title{Normalize Type Strings}
\usage{
normalize_type_strings(type_strings, types)
}
\description{
Normalize Type Strings
}
\examples{
types <- get_ff_types()
type_strings <- c("Apple")
normalize_type_strings("Apple", types)
normalize_type_strings(c("14", "[Malus domestica]"), types)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{parse_type_strings}}
}

