% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{normalize_type_strings}
\alias{normalize_type_strings}
\title{Normalize Type Strings}
\usage{
normalize_type_strings(type_strings, types = get_ff_types(pending = FALSE,
  urls = FALSE))
}
\arguments{
\item{type_strings}{Character vector of type strings as comma-delimited lists.}

\item{types}{Falling Fruit types}
}
\value{
Character vector of complete type strings as comma-delimited lists.
}
\description{
Checks type strings against Falling Fruit types and fills in any missing components.
}
\examples{
ff_types <- get_ff_types()
normalize_type_strings("Apple", ff_types)
normalize_type_strings(c("14", "Apple [Malus]"), ff_types)
normalize_type_strings(c("", " ,", NA), ff_types)
normalize_type_strings(c("14: Apple, 14: Apple"), ff_types)
normalize_type_strings("Hello World", ff_types)
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_descriptions}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{parse_type_strings}}
}
