% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{parse_col_search}
\alias{parse_col_search}
\title{Parse Catalogue of Life (COL) Search Results}
\usage{
parse_col_search(search, types = c("results", "ids"), exact = TRUE,
  scientific_name = TRUE, accepted_name = TRUE, ignore.case = TRUE)
}
\description{
Parse Catalogue of Life (COL) Search Results
}
\examples{
s <- get_col_search("Malus domestica")
str(parse_col_search(s))
str(parse_col_search(s, accepted_name = FALSE))
str(parse_col_search(s, exact = FALSE))
s <- get_col_search("Apple")
str(parse_col_search(s, scientific_name = FALSE))
}
\seealso{
Other COL functions: \code{\link{build_col_scientific_name}},
  \code{\link{get_col_page}}, \code{\link{get_col_search}},
  \code{\link{parse_col_page}}
}

