% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingfruit.R
\name{parse_type_strings}
\alias{parse_type_strings}
\title{Parse Type Strings}
\usage{
parse_type_strings(type_strings)
}
\description{
Parse Type Strings
}
\examples{
type_strings <- build_type_strings(c(1, 2), c("Apple", "Pear"), c("Malus domestica", "Pyrus communis"))
str(parse_type_strings(type_strings))
}
\seealso{
Other Falling Fruit functions: \code{\link{build_location_description}},
  \code{\link{build_type_strings}},
  \code{\link{expand_category_mask}},
  \code{\link{get_ff_types}},
  \code{\link{match_type_strings}},
  \code{\link{normalize_type_strings}}
}

