% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wikicommons_page}
\alias{parse_wikicommons_page}
\title{Parse Wikimedia Commons Page}
\usage{
parse_wikicommons_page(page, types = c("common_names"))
}
\arguments{
\item{page}{(response) Result of \code{\link{get_wiki_page}}.}

\item{types}{(character) Data types to parse.}
}
\description{
Parse Wikimedia Commons Page
}
\examples{
pg <- get_wiki_page("https://commons.wikimedia.org/wiki/Malus_domestica")
str(parse_wikicommons_page(pg))
}
\seealso{
Other Wiki functions: \code{\link{build_wiki_url}},
  \code{\link{get_wiki_page}},
  \code{\link{parse_wiki_url}},
  \code{\link{parse_wikipedia_page}},
  \code{\link{parse_wikispecies_page}}
}
