% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{query_sources_about_type}
\alias{query_sources_about_type}
\title{Query Sources about Falling Fruit (FF) Type}
\usage{
query_sources_about_type(ff_type, sources = c("eol", "col", "inaturalist",
  "wikipedia", "wikicommons", "wikispecies"))
}
\description{
Query Sources about Falling Fruit (FF) Type
}
\examples{
ff_types <- get_ff_types(urls = TRUE)
query_sources_about_type(ff_types[id == 14], sources = "eol")
}

