% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{read_locations}
\alias{read_locations}
\title{Load Locations Dataset}
\usage{
read_locations(file, xy = c("lng", "lat"), id = "id",
  CRSobj = sp::CRS("+proj=longlat +ellps=WGS84"), stringsAsFactors = FALSE,
  na.strings = c("", "NA", "N/A", "na", "n/a"), ...)
}
\arguments{
\item{file}{The path of the file to be read.}

\item{xy}{Names of the x and y coordinate fields (renamed to "lng", "lat" respectively).}

\item{id}{Name of the id field (renamed to "id").}

\item{CRSobj}{Coordinate reference system (\code{\link{sp::CRS}}).}

\item{...}{Additional parameters passed to \code{\link{data.table::fread}} (delimited files), \code{\link{rgdal::readOGR}} (spatial data), or \code{\link{xml2::read_xml}} (certain kml files).}
}
\description{
Load Locations Dataset
}
\seealso{
Other location import functions: \code{\link{apply_match_table}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{merge_overlapping_locations}},
  \code{\link{write_locations_for_import}}
}
