% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{read_locations}
\alias{read_locations}
\title{Load Locations Dataset}
\usage{
read_locations(file, id = NULL, xy = c("lng", "lat"), proj4 = NULL, ...)
}
\arguments{
\item{file}{The path of the file to be read.}

\item{id}{Name of the column with a unique identifier (renamed to "id"). If \code{NULL}, locations are assigned an integer id corresponding to their order of appearance in the file.}

\item{xy}{(delimited text files only) Names of the columns with x and y coordinates (renamed to "lng", "lat" respectively).}

\item{proj4}{Current coordinate reference system expressed as a proj.4 string (\url{http://proj4.org/parameters.html}) or EPSG integer code (\url{http://spatialreference.org/ref/epsg/}). If specified, overrides any embedded value (spatial data). If \code{NULL}, the embedded value is used (spatial data) or WGS 84 (EPSG:4326) is assumed if missing.}

\item{...}{Additional parameters passed to \code{\link[data.table]{fread}} (delimited text files), \code{\link[rgdal]{readOGR}} (spatial data), or \code{\link[xml2]{read_xml}} (certain kml files).}
}
\description{
Load Locations Dataset
}
\seealso{
Other location import functions: \code{\link{apply_match_table}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{merge_overlapping_locations}},
  \code{\link{write_locations_for_import}}
}
