% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_values_in_list}
\alias{replace_values_in_list}
\title{Replace Values in List}
\usage{
replace_values_in_list(x, old, new)
}
\arguments{
\item{x}{List object.}

\item{old}{List or vector of values to replace.}

\item{new}{List or vector of replacement values.}
}
\value{
A list with values replaced.
}
\description{
Iterates through a list and replaces values with replacements.
}
\examples{
str(x <- list(NULL, list(NULL, NA, 2)))
str(replace_values_in_list(x, NULL, 1))
str(replace_values_in_list(x, list(NULL, NA), 1))
str(replace_values_in_list(x, list(NULL, NA), list(1, 2)))
}
\seealso{
Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{lvapply}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}}, \code{\link{sptransform}},
  \code{\link{subset_search_results}},
  \code{\link{unique_na}}
}
