% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unique_na}
\alias{unique_na}
\title{Return Single Unique Value or NA}
\usage{
unique_na(x, na.rm = FALSE)
}
\arguments{
\item{x}{List or atomic vector.}

\item{na.rm}{Whether missing values should be removed.}
}
\description{
Returns either \code{NA} or the single unique element if it exists.
}
\examples{
unique_na(c(1, 1))
unique_na(c(1, 2))
unique_na(c(1, NA), na.rm = FALSE)
unique_na(c(1, NA), na.rm = TRUE)
}
\seealso{
Other helper functions: \code{\link{capitalize_words}},
  \code{\link{ch1903_to_wgs84}},
  \code{\link{clean_strings}},
  \code{\link{format_addresses}},
  \code{\link{format_scientific_names}},
  \code{\link{is.empty}}, \code{\link{lvapply}},
  \code{\link{melt_by_listcol}}, \code{\link{na.remove}},
  \code{\link{normalize_language}},
  \code{\link{parse_scientific_names}},
  \code{\link{quotemeta}},
  \code{\link{replace_values_in_list}},
  \code{\link{sp_transform}},
  \code{\link{subset_search_results}}
}
