% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{write_locations_for_import}
\alias{write_locations_for_import}
\title{Write Locations to File for Import}
\usage{
write_locations_for_import(dt, file, drop_extra_fields = TRUE)
}
\description{
See http://fallingfruit.org/locations/import for format.
FIXME: Do not edit dt in place.
}
\seealso{
Other location import functions: \code{\link{apply_ff_type_matches}},
  \code{\link{match_to_ff_types}},
  \code{\link{read_locations}}
}

