% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{write_locations_for_import}
\alias{write_locations_for_import}
\title{Write Locations to File for Import}
\usage{
write_locations_for_import(dt, file)
}
\arguments{
\item{dt}{Locations data.}

\item{file}{Path to file.}
}
\description{
Write Locations to File for Import
}
\seealso{
Other location import functions: \code{\link{apply_match_table}},
  \code{\link{build_match_table}},
  \code{\link{match_names_to_ff_types}},
  \code{\link{merge_overlapping_locations}},
  \code{\link{read_locations}}
}
